fix(transition): cast progress to double in ImageTableDrawFrame log output
Ensure proper format when logging out-of-range progress values in C 
transition renderer by explicitly casting to double. Prevents warnings 
during build.

+++ core/transitions/roxy_transition.c
@@ -144,7 +144,8 @@ int roxy_transition_imageTableDrawFrame_l(lua_State* L)
 
     // Validate progress range
     if (progress < 0.0f || progress > 1.0f) {
-        pd->system->logToConsole("Roxy ERROR: ImageTableDrawFrame - progress must be in range [0.0, 1.0], got %f", progress);
+        pd->system->logToConsole("Roxy ERROR: ImageTableDrawFrame - progress must be in range [0.0, 1.0], got %f", (double)progress);
+
         return 0;
     }
